preprocess/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
mysql-connector-python==8.3.0
scikit-learn==1.4.0
pandas==2.2.1
streamlit==1.22.0